pyproject.toml
[project]
name = "adaptation-atlas-co-pilot"
version = "0.1.0"
description = "Adaptation Atlas Co-Pilot"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langgraph>=0.6.1",
    "langchain-mistralai>=0.2.12",
    "dotenv>=0.9.9",
    "pydantic-settings>=2.11.0",
    "langchain-chroma>=0.2.6",
    "mistralai>=1.9.10",
    "google-genai>=1.39.1",
    "pyarrow>=15.0.0",
    "pandas>=2.0.0",
    "duckdb>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["aacp*"]

[tool.uv.sources]
my_base_project = { path = "adaptation-atlas-co-pilot" }

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pre-commit>=3.0.0",
    "ruff>=0.1.0",
    "langchain-chroma>=0.2.6",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
]
